pyproject.toml
[tool.poetry]
name = "churn-mlops-project"
version = "0.1.0"
description = "End-to-end MLOps pipeline for predicting customer churn using e-commerce transaction data. Includes data processing, model training, evaluation, CI/CD, and deployment."
authors = ["Leonel481 <leonel.aliaga.v@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
pandas = ">=2.3.0,<3.0.0"
scikit-learn = ">=1.7.0,<2.0.0"
xgboost = ">=3.0.2,<4.0.0"
shap = ">=0.48.0,<0.49.0"
hyperopt = ">=0.2.7,<0.3.0"
setuptools = ">=80.9.0,<81.0.0"
lightgbm = ">=4.6.0,<5.0.0"
google-cloud-storage = ">=3.2.0,<4.0.0"
google-cloud-bigquery = ">=3.35.0,<4.0.0"
pyarrow = ">=21.0.0,<22.0.0"
gcsfs = ">=2025.7.0,<2026.0.0"
prefect = {version = ">=3.4.9,<4.0.0", extras = ["docker"]}
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
requests = "^2.32.4"
ipykernel = "^6.29.5"
matplotlib = "^3.10.3"
seaborn = "^0.13.2"
s3fs = "^2025.7.0"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

